Rename HTML agent output blog
